requirements.txt
praw~=7.8.1
python-dotenv~=1.0.1

matplotlib~=3.10.0

protobuf~=5.29.3
pandas~=2.2.3
plotly~=5.24.1
streamlit~=1.41.1

altair~=5.5.0
